[task] 주문 목록 페이지 제거 및 당첨자 페이지에 메인화면으로 가기 버튼 추가 (#1032)
