+++ frontend/src/components/@common/GameActionButton/GameActionButton.stories.tsx
@@ -32,7 +32,7 @@ export const CardGame: Story = {
     isDisabled: false,
     gameName: '카드게임',
     description: ['2라운드 동안 매번 카드 1장씩 뒤집어', '가장 높은 점수를 내보세요!'],
-    icon: <img src={CardIcon} alt="card game" />,
+    icon: <img src={CardIcon} alt="card game" style={{ width: '30px', height: 'auto' }} />,
   },
 };
 
@@ -42,7 +42,7 @@ export const CardGameSelected: Story = {
     isDisabled: false,
     gameName: '카드게임',
     description: ['2라운드 동안 매번 카드 1장씩 뒤집어', '가장 높은 점수를 내보세요!'],
-    icon: <img src={CardIcon} alt="card game" />,
+    icon: <img src={CardIcon} alt="card game" style={{ width: '30px', height: 'auto' }} />,
   },
 };
 
@@ -52,7 +52,7 @@ export const RacingGame: Story = {
     isDisabled: false,
     gameName: '레이싱게임',
     description: ['화면을 클릭해 속도를 높여서', '가장 먼저 도착하세요!'],
-    icon: <img src={RacingIcon} alt="racing game" />,
+    icon: <img src={RacingIcon} alt="racing game" style={{ width: '30px', height: 'auto' }} />,
   },
 };
 
@@ -62,7 +62,7 @@ export const RacingGameSelected: Story = {
     isDisabled: false,
     gameName: '레이싱게임',
     description: ['화면을 클릭해 속도를 높여서', '가장 먼저 도착하세요!'],
-    icon: <img src={RacingIcon} alt="racing game" />,
+    icon: <img src={RacingIcon} alt="racing game" style={{ width: '30px', height: 'auto' }} />,
   },
 };
 
@@ -72,6 +72,6 @@ export const Disabled: Story = {
     isDisabled: true,
     gameName: '카드게임',
     description: ['2라운드 동안 매번 카드 1장씩 뒤집어', '가장 높은 점수를 내보세요!'],
-    icon: <img src={CardIcon} alt="card game" />,
+    icon: <img src={CardIcon} alt="card game" style={{ width: '30px', height: 'auto' }} />,
   },
 };
