--- frontend/src/features/room/order/components/PlayerMenu/PlayerMenu.styled.ts
@@ -1,13 +1,0 @@
-import styled from '@emotion/styled';
-
-export const OrderList = styled.section`
-  height: calc(100% - 4rem);
-  overflow-y: scroll;
-`;
-
-export const OrderItem = styled.article`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.75rem 0;
-`;
