--- frontend/src/features/room/order/pages/OrderPage.styled.ts
@@ -1,23 +1,0 @@
-import styled from '@emotion/styled';
-
-export const BannerContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  gap: 5px;
-`;
-
-export const Logo = styled.img`
-  width: 100px;
-  height: 50px;
-  object-fit: contain;
-`;
-
-export const ListHeader = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 1rem;
-`;
