+++ frontend/src/features/room/roulette/pages/RoulettePlayPage/hooks/useRoulettePlay.ts
@@ -1,12 +1,12 @@
 import { useWebSocket } from '@/apis/websocket/contexts/WebSocketContext';
 import { useIdentifier } from '@/contexts/Identifier/IdentifierContext';
+import { useReplaceNavigate } from '@/hooks/useReplaceNavigate';
 import { RouletteWinnerResponse } from '@/types/roulette';
 import { useEffect, useState } from 'react';
-import { useReplaceNavigate } from '@/hooks/useReplaceNavigate';
 
 const useRoulettePlay = () => {
   const { joinCode, myName } = useIdentifier();
-  const { send } = useWebSocket();
+  const { send, stopSocket } = useWebSocket();
   const navigate = useReplaceNavigate();
   const [winner, setWinner] = useState<string | null>(null);
   const [randomAngle, setRandomAngle] = useState(0);
@@ -23,16 +23,19 @@ const useRoulettePlay = () => {
   };
 
   useEffect(() => {
-    // TODO: 당첨자가 나오지 않았을 때, 에러 처리 방식 정하기
-    if (!winner || !winner.trim()) console.warn('당첨자가 추첨되지 않았습니다.');
+    if (!isSpinStarted) return;
 
-    if (isSpinStarted) {
-      const timer = setTimeout(() => {
-        navigate(`/room/${joinCode}/roulette/result`, { state: { winner } });
-      }, 5000);
-      return () => clearTimeout(timer);
+    if (!winner || !winner.trim()) {
+      console.warn('당첨자가 추첨되지 않았습니다.');
     }
-  }, [isSpinStarted, winner, navigate, joinCode]);
+
+    const timer = setTimeout(() => {
+      stopSocket();
+      navigate(`/room/${joinCode}/roulette/result`, { state: { winner } });
+    }, 5000);
+
+    return () => clearTimeout(timer);
+  }, [isSpinStarted, winner, navigate, joinCode, stopSocket]);
 
   return {
     winner,
